+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # REPLACE_GLOBALLY_WITH_NEXT_VERSION
 - PPI-174 - Cart and order line items are now sent with their SKU
 - PPI-174 - Added events to adjust line items which are sent to PayPal
+- PPI-202 - Fix PayPal checkout for customers with net prices
 
 # 2.0.2
 - PPI-199 - Improve webhook log messages